refactor(cli): give startup planning its own module
Move the planning machinery out of simnet/mod.rs into simnet/startup.rs:
the runbook execution mode, the watchdog, the seal plumbing, and the plan
assembly and dispatch. DeploymentPlan becomes StartupPlan and
plan_deployment becomes plan_startup, because the plan carries every
startup task planning finds: clones, deployments, and whatever arrives
next. The execution thread is named Startup Runbook Executions for the
same reason.

Working here surfaced a few lints. CheckVersionResponse and the serde
imports exist only for the version_check feature and are now gated with
it; log::info is gated to Linux, where its one caller (the daemon fork)
lives; the tests module moves to the end of simnet/mod.rs so no items
follow it; assemble_runbook_execution_futures takes a slice and reuses
its own type alias.
